Time offset error and in memory database
Time offset:
The SQLite database stores dates in UTC whilst node Date object uses local time. Conversion between these must be manually handled. This is done by having functions that
 - converts from databases UTC time string to local time Date() object and
 - converts from local time Date() object to UTC time string.
A new interface that has time stored as string is therefore introduced.

In memory database:
Added option to run database in memory. This still requires a database on disk, but this is then cloned into memory and thus not changed.